+++ lib/src/main/java/org/asamk/signal/manager/api/Device.java
@@ -1,3 +1,3 @@
 package org.asamk.signal.manager.api;
 
-public record Device(long id, String name, long created, long lastSeen, boolean isThisDevice) {}
+public record Device(int id, String name, long created, long lastSeen, boolean isThisDevice) {}
